+++ test/smokes/scss_lint/expectations.rb
@@ -33,7 +33,7 @@ Smoke.add_test("success", {
     ],
     analyzer: {
       name: 'SCSS-Lint',
-      version: '0.58.0'
+      version: '0.59.0'
     }
 })
 
@@ -45,7 +45,7 @@ Smoke.add_test("no_scss_files", {
     ],
     analyzer: {
       name: 'SCSS-Lint',
-      version: '0.58.0'
+      version: '0.59.0'
     }
 }, warnings: [
   {message: 'No files, paths, or patterns were specified', file: nil}
@@ -78,7 +78,7 @@ Smoke.add_test("with_config_option", {
     ],
     analyzer: {
       name: 'SCSS-Lint',
-      version: '0.58.0'
+      version: '0.59.0'
     }
 })
 
@@ -96,7 +96,7 @@ Smoke.add_test("syntax_error", {
     ],
     analyzer: {
       name: 'SCSS-Lint',
-      version: '0.58.0'
+      version: '0.59.0'
     }
 })