+++ test/smokes/code_sniffer/expectations.rb
@@ -15,7 +15,7 @@ Smoke.add_test("success", {
   ],
   analyzer: {
     name: "code_sniffer",
-    version: "3.5.0"
+    version: "3.5.2"
   },
 })
 
@@ -34,7 +34,7 @@ Smoke.add_test("specified_dir", {
   ],
   analyzer: {
     name: "code_sniffer",
-    version: "3.5.0"
+    version: "3.5.2"
   },
 }, warnings: [{ :message => "`dir` key under the `options` is deprecated. Please declare it just under the `code_sniffer`. See https://help.sider.review/tools/php/codesniffer#options", :file => "sideci.yml" }])
 
@@ -47,7 +47,7 @@ Smoke.add_test("sideci_php_sandbox", {
     issues: :_,
     analyzer: {
         name: "code_sniffer",
-        version: "3.5.0"
+        version: "3.5.2"
     },
 })
 
@@ -58,7 +58,7 @@ Smoke.add_test("with_php_version", {
   issues: :_,
   analyzer: {
     name: "code_sniffer",
-    version: "3.5.0"
+    version: "3.5.2"
   }
 })
 
@@ -85,7 +85,7 @@ Smoke.add_test("version_2", {
   ],
   analyzer: {
     name: "code_sniffer",
-    version: "3.5.0"
+    version: "3.5.2"
   },
 },
 warnings: [{
@@ -108,7 +108,7 @@ Smoke.add_test("autodetect_cakephp", {
   ],
   analyzer: {
     name: "code_sniffer",
-    version: "3.5.0"
+    version: "3.5.2"
   },
 })
 
@@ -134,6 +134,6 @@ Smoke.add_test("autodetect_symfony", {
   ],
   analyzer: {
     name: "code_sniffer",
-    version: "3.5.0"
+    version: "3.5.2"
   },
 })