+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 - [RuboCop] Empower runner [#472](https://github.com/sider/runners/pull/472)
 - [checkstyle] Empower runner [#475](https://github.com/sider/runners/pull/475)
 - [SwiftLint] Show a warning when no linting files [#478](https://github.com/sider/runners/pull/478)
+- Bump JavaSee from 0.1.2 to 0.1.3 [#479](https://github.com/sider/runners/pull/479)
 
 ## 0.8.2