+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-[Full diff](https://github.com/sider/runners/compare/0.8.2...HEAD)
+[Full diff](https://github.com/sider/runners/compare/0.9.0...HEAD)
+
+## 0.9.0
+
+[Full diff](https://github.com/sider/runners/compare/0.8.2...0.9.0)
 
 - Update scss_lint from 0.58.0 to 0.59.0 [#432](https://github.com/sider/runners/pull/432)
 - Bump pmd_java from 6.18.0 to 6.19.0 [#433](https://github.com/sider/runners/pull/433)